totalPages type changed to integer in ResponseBody

+++ src/main/java/org/khu/structure/ResponseBody.java
@@ -18,7 +18,7 @@ public class ResponseBody {
 
     private Long totalElements;
 
-    private Long totalPages;
+    private Integer totalPages;
 
     private String message;
